feat: extract compile path
