+++ .changeset/smart-ties-peel.md
@@ -1,0 +1,5 @@
+---
+"@wbe/low-router": minor
+---
+
+extract compilePath static method as external function
