+++ packages/low-router/src/index.ts
@@ -1,6 +1,6 @@
 export { LowRouter } from "./LowRouter"
 export type { RouteContext, Route, RouteParams, RouteProps, RouterOptions } from "./types"
-export { createMatcher, pathToRegexp } from "./createMatcher"
+export { createMatcher, pathToRegexp } from "./utils/createMatcher"
 
 // additional exports
 export { createBrowserHistory } from "./utils/browserHistory"
